audio: use prebuilt HAL for now [2/2]
Motorola made some modifications to the audio HAL that I haven't
yet figured out that result in changes to microphone operation in
speakerphone mode and other things. The stock qcom HAL in CyanogenMod
doesn't have these changes, and thus some audio functionality is
broken. In the long term, I should figure out what Motorola changed
and make it work with the source built HAL, but for now I just want
it to work, so I'll use the Motorola binary for now.

The other commit [1/2] is in the device tree.